refactor: 검증책임 분리
