+++ src/main/java/lotto/GameProcess/LottoProcessImpl.java
@@ -6,6 +6,7 @@ import java.util.List;
 import lotto.Input.InputNumber;
 import lotto.Input.InputNumberImpl;
 import lotto.dataObjects.Lotto;
+import lotto.dataObjects.Money;
 import lotto.utils.LottoValidator;
 import lotto.utils.LottoValidatorImpl;
 
@@ -20,9 +21,9 @@ public class LottoProcessImpl implements LottoProcess {
 
 	@Override
 	public void playGame() {
-		//데이터/로직 담은 객체 생성
 		int lottoPrice = inputNumber.inputPurchasePrice();
-		
+		Money money = new Money(lottoPrice);
+
 		List<Integer> lottoNumbers = new ArrayList<>();
 		inputNumber.inputWinningNumbers(lottoNumbers);
 		Lotto winningLottoNumbers = new Lotto(lottoNumbers);
