+++ src/main/java/lotto/Input/InputNumber.java
@@ -2,8 +2,6 @@ package lotto.Input;
 
 import java.util.List;
 
-import lotto.dataObjects.Lotto;
-
 public interface InputNumber {
 
 	int inputPurchasePrice();
